# contracts/openapi.yml
openapi: 3.0.3
info:
  title: Food Commerce API
  description: API First contract for modular monolithic food e-commerce
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /inventory/{productId}:
    get:
      tags: [Inventory]
      summary: Get inventory status
      operationId: getInventory
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
...
  /loyalty/{customerId}:
    get:
      tags: [Loyalty]
      summary: Get loyalty points
      operationId: getLoyalty
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyAccount'
...
  /admin/inventory/{productId}/adjust:
    patch:
      tags: [Inventory]
      summary: Adjust inventory manually (admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
      responses:
        '200':
          description: OK
...
  /admin/loyalty/{customerId}/adjust:
    patch:
      tags: [Loyalty]
      summary: Adjust loyalty points manually
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [points]
              properties:
                points:
                  type: integer
      responses:
        '200':
          description: Loyalty updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyAccount'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Inventory:
      type: object
      required: [productId, availableQuantity, reservedQuantity]
      properties:
        productId:
          type: string
        availableQuantity:
          type: integer
          minimum: 0
        reservedQuantity:
          type: integer
          minimum: 0
...
    LoyaltyAccount:
      type: object
      required: [customerId, points]
      properties:
        customerId:
          type: string
        points:
          type: integer
          minimum: 0
security:
  - bearerAuth: []